Reuse connection
Tap-postgres opens a new connection every time it needs to cast a value.
This is highly inefficient as opening a connection is usually a slow and
resource-intensive operation. An easy fix would be to use something like
PgBouncer, but it's even better if we open just once connection and
reuse it for all queries.

An easy way to fix that is to implement Singleton, but we decided to go
even simpler and store our connections in the db module we already have.
All we do is to store two types of connections (transactional and
logical). Every time we need a connection we check if we already have an
open connection. If we don't we open it and store it. Then just return
that connection.