Improve logging message
Changed from debug -> info level. I deployed the changes from this PR,
and I wanted to make sure things were working as expected. It was in an
environment with sensitive data, so I wasn't able to simply switch to
debug level since that would lead the sensitive data to the logs. Also
added the column being affected to the message.