Enforce ordering by replication key during full refresh
For the incremental strategy if we stop an initial data load midway our state will be that of the last record that was written to the target. Right now the generated SQL for this case does not enforce ordering records by the replication_key. If we don't enforce the ordering of the records from the source we have no guarantee that all records before the last one were written to the target (as Postgres might return them in disorder). To overcome these issues we should enforce ordering.